# src/main/resources/api.yaml
openapi: "3.0.3"
info:
  title: Biblioteca API
  description: API para gerenciar organizações
  version: 1.0.0
...
paths:
...
  /usuario/find:
    post:
      summary: Busca usuários
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: findUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequest'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'
  /usuario/detail/{id}:
    get:
      summary: Detalha  um usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: detailUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'
  /usuario/create:
    put:
      summary: Cria um usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: createUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioModel'
  /usuario/update:
    patch:
      summary: Atualiza um usuário
      security:
          - bearerAuth: []
      tags:
        - Usuario
      operationId: updateUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioModel'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioModel'
  /usuario/delete/{id}:
    delete:
      summary: Deleta um usuário
      security:
          - bearerAuth: []
      tags:
        - Usuario
      operationId: deleteUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # Especifica o formato do token, pode ser omitido
  schemas:
...
    UsuarioRequest:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        dataNascimento:
          type: string
          format: date
        perfis:
          type: array
          items:
            type: integer
        limit:
            type: integer
        offset:
          type: integer
        event:
          type: object
          $ref: '#/components/schemas/EventPagination'
    EventPagination:
      type: object
      properties:
        length:
          type: integer
        pageIndex:
          type: integer
        pageSize:
            type: integer
        previousPageIndex:
            type: integer
...
    UsuarioResponseDTO:
      type: object
      properties:
        usuarios:
          type: array
          items:
            $ref: '#/components/schemas/UsuarioModel'
        totalRecords:
          type: integer
...
    RegisterRequestDTO:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        senha:
          type: string
        dataNascimento:
          type: string
          format: date
        enderecos:
          type: array
          items:
            $ref: '#/components/schemas/EnderecoModel'
        perfis:
          type: array
          items:
            $ref: '#/components/schemas/PerfilModel'
...
    UsuarioModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        senha:
          type: string
        email:
          type: string
        dataNascimento:
          type: string
        enderecos:
          type: array
          items:
            $ref: '#/components/schemas/EnderecoModel'
        perfis:
            type: array
            items:
                $ref: '#/components/schemas/PerfilModel'
    EnderecoModel:
      type: object
      properties:
        id:
          type: integer
        logradouro:
          type: string
        numero:
          type: integer
        complemento:
            type: string
        bairro:
          type: string
        cidade:
          type: string
        uf:
          $ref: '#/components/schemas/UnidadeFederativaModel'
        cep:
          type: string
...
    PerfilModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
    UnidadeFederativaModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sigla:
            type: string
...